characters in password can not be escaped.
