+++ src/webhook/validation/activegate.go
@@ -1,15 +1,10 @@
 package validation
 
 import (
-	"context"
 	"fmt"
-	"net/url"
 
-	"github.com/Dynatrace/dynatrace-operator/src/agproxysecret"
 	dynatracev1beta1 "github.com/Dynatrace/dynatrace-operator/src/api/v1beta1"
 	corev1 "k8s.io/api/core/v1"
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
 const (
@@ -24,16 +19,6 @@ Make sure you correctly specify the ActiveGate capabilities in your custom resou
 Make sure you don't duplicate an Activegate capability in your custom resource.
 `
 	warningMissingActiveGateMemoryLimit = `ActiveGate specification missing memory limits. Can cause excess memory usage.`
-
-	errorInvalidActiveGateProxyUrl = `The DynaKube's specification has an invalid Proxy URL value set. Make sure you correctly specify the URL in your custom resource.`
-	errorInvalidEvalCharacter      = `The DynaKube's specification has an invalid Proxy password value set. Make sure you correctly escape quotation mark, backtick and backslash characters using backslash.`
-
-	errorMissingActiveGateProxySecret = `The Proxy secret indicated by the DynaKube specification doesn't exist.`
-
-	errorInvalidProxySecretFormat = `The Proxy secret indicated by the DynaKube specification has an invalid format. Make sure you correctly creates the secret.`
-
-	errorInvalidProxySecretUrl           = `The Proxy secret indicated by the DynaKube specification has an invalid URL value set. Make sure you correctly specify the URL in the secret.`
-	errorInvalidProxySecretEvalCharacter = `The Proxy secret indicated by the DynaKube specification has an invalid Proxy password value set. Make sure you correctly escape quotation mark, backtick and backslash characters using backslash.`
 )
 
 func conflictingActiveGateConfiguration(dv *dynakubeValidator, dynakube *dynatracev1beta1.DynaKube) string {
@@ -81,65 +66,6 @@ func missingActiveGateMemoryLimit(dv *dynakubeValidator, dynakube *dynatracev1be
 	return ""
 }
 
-func invalidActiveGateProxyUrl(dv *dynakubeValidator, dynakube *dynatracev1beta1.DynaKube) string {
-	if dynakube.Spec.Proxy != nil {
-		if len(dynakube.Spec.Proxy.ValueFrom) > 0 {
-			var proxySecret corev1.Secret
-			err := dv.clt.Get(context.TODO(), client.ObjectKey{Name: dynakube.Spec.Proxy.ValueFrom, Namespace: dynakube.Namespace}, &proxySecret)
-			if k8serrors.IsNotFound(err) {
-				return errorMissingActiveGateProxySecret
-			} else if err != nil {
-				return fmt.Sprintf("error occurred while reading PROXY secret indicated in the Dynakube specification (%s)", err.Error())
-			}
-			proxyUrl, ok := proxySecret.Data[agproxysecret.ProxySecretKey]
-			if !ok {
-				return errorInvalidProxySecretFormat
-			}
-			return validateProxyUrl(string(proxyUrl), errorInvalidProxySecretUrl, errorInvalidProxySecretEvalCharacter)
-		} else if len(dynakube.Spec.Proxy.Value) > 0 {
-			return validateProxyUrl(dynakube.Spec.Proxy.Value, errorInvalidActiveGateProxyUrl, errorInvalidEvalCharacter)
-		}
-	}
-	return ""
-}
-
 func memoryLimitSet(resources corev1.ResourceRequirements) bool {
 	return resources.Limits != nil && resources.Limits.Memory() != nil
-}
-
-// proxyUrl is valid if
-// 1) encoded
-// 2) "`\ are escaped using \
-func validateProxyUrl(proxyUrl string, parseErrorMessage string, evalErrorMessage string) string {
-	if parsedUrl, err := url.Parse(proxyUrl); err != nil {
-		return parseErrorMessage
-	} else {
-		password, _ := parsedUrl.User.Password()
-		if isEvalEscapeNeeded(password) {
-			return evalErrorMessage
-		}
-	}
-	return ""
-}
-
-// 'eval' command is used by entrypoint.sh:readSecret function to return its result.
-// For this reason quotation mark ", backtick ` and backslash \ characters have to be escaped using backslash.
-func isEvalEscapeNeeded(str string) bool {
-	previousChar := '\000'
-	for _, char := range str {
-		if char == '"' || char == '`' {
-			if previousChar != '\\' {
-				return true
-			}
-			previousChar = char
-		} else if previousChar == '\\' {
-			if char != '\\' {
-				return true
-			}
-			previousChar = '\000'
-		} else {
-			previousChar = char
-		}
-	}
-	return false
 }
